Move structure related utility methods to utils/structure.js.
