+++ packages/ckeditor5-table/src/commands/setheadercolumncommand.js
@@ -11,12 +11,11 @@ import Command from '@ckeditor/ckeditor5-core/src/command';
 
 import {
 	findAncestor,
-	getHorizontallyOverlappingCells,
 	isHeadingColumnCell,
-	splitVertically,
 	updateNumericAttribute
 } from '../utils/common';
 import { getColumnIndexes, getSelectionAffectedTableCells } from '../utils/selection';
+import { getHorizontallyOverlappingCells, splitVertically } from '../utils/structure';
 
 /**
  * The header column command.
