+++ packages/ckeditor5-table/src/tableclipboard.js
@@ -12,14 +12,18 @@ import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
 import TableSelection from './tableselection';
 import TableWalker from './tablewalker';
 import {
-	getVerticallyOverlappingCells,
-	getHorizontallyOverlappingCells,
-	splitHorizontally,
-	splitVertically, findAncestor
+	findAncestor
 } from './utils/common';
-import { cropTableToDimensions, trimTableCellIfNeeded } from './tableselection/croptable';
 import TableUtils from './tableutils';
 import { getColumnIndexes, getRowIndexes, getSelectionAffectedTableCells, isSelectionRectangular } from './utils/selection';
+import {
+	cropTableToDimensions,
+	getHorizontallyOverlappingCells,
+	getVerticallyOverlappingCells,
+	splitHorizontally,
+	splitVertically,
+	trimTableCellIfNeeded
+} from './utils/structure';
 
 /**
  * This plugin adds support for copying/cutting/pasting fragments of tables.
